kernel: Create kernel-devel squashfs and bind into host containers
As part of the kernel build install the kernel development headers into
a squashfs file installed onto the Thar host. On boot this will be
unpacked to make the headers available, saving a significant amount of
space in the root partition.

For superpowered host containers host-ctr bind mounts in
	/lib/modules -> /lib/modules, and
	/usr/src/kernels -> /usr/src/kernels
allowing the kernel headers to be transparently accessed from within the
container.

Signed-off-by: Samuel Mendoza-Jonas <samjonas@amazon.com>