Replace unused timezone/hostname settings with motd setting
It's useful to have an example setting so that users can see it in
documentation and follow along, learning the API by actually making calls,
without worrying about breaking their system.  We were using hostname and
timezone for this, but those sounded like real settings (and were originally
intended to be real settings) so it was confusing.

motd is a "real" setting in that it updates /etc/motd, but that's very low risk
and only exposed to the user in any case.  Plus, with low risk, it's fun to see
an actual impact on your system.