Exclude README.md from cargo change tracking
In packages using cargo-readme, README.md files are touched by build.rs, which
triggers cargo's change tracking in the next build.  This means cargo is
rebuilding the package every time.

This change adds README.md to the exclude list in Cargo.toml files so that
cargo doesn't think a README change requires a rebuild.  Most of our packages
use cargo-readme to generate README.md, and for packages that don't, this still
isn't harmful; manual updates to README.md don't need a Rust build.  (Special
cases can just remove this exclude line.)

The downside is that accidental changes to README.md, like removing it, won't
be automatically corrected with a rebuild if you haven't touched anything else.
This should be rare, and will correct itself upon any other change.  (And a
developer is unlikely to commit such an accidental change.)