pubsys: only create tokio runtime when needed
tough-ssm creates its own tokio runtime for making SSM calls, so we can't use
it inside our own tokio runtime.  This change removes the general
`[tokio::main]` annotation that creates a runtime for the entire app, and
instead only creates a runtime inside the `ami` subcommand that needs one.  We
can switch back to the annotation when tough-ssm moves to an async interface.