tough: update to v0.10.0
This version of tough changes the interface of the Repository struct
and its load function. The usage sites have been updated to use the
new interface.

updog: The query parameter mechanism needed to be changed to
accommodate the fact that Repository now owns a Transport instead of
holding a reference. This is done by holding an Arc pointer to the
query params both inside and outside of the custom Transport.

pubsys: A custom Transport is no-longer necessary (its purpose was to
support both http and file transport, which DefaultTransport now does).
Additionally, the new TransportError gives pubsys the information it
needs to know whether a repo already exists or not.

update_metadata and migrator: no significant changes.