kernel: overhaul kernel development sources
Previously, we included host programs like `objtool` which are built
with the default `gcc` compiler and not our cross-compiler toolchain.

This works as long as the running system matches our build host, but
would break if we began building x86_64 images on an aarch64 system.
The reverse is not true today, but only because `objtool` is not yet
required for the arm64 target.

Ideally, we'd be able to cross-compile these host programs, but that
isn't supported by the kernel's build system, and would be hard to
implement. For example, `fixdep` is both a tool we'd want to ship,
meaning it would need to be cross-compiled, and a tool that's used
to build `objtool`, meaning it couldn't be cross-compiled and still
run on the build host.

Instead we push the problem out to the downstream consumer, who can
be relied on to have a compiler that can build native versions of the
host programs. This requires shipping all the headers, tools, and
scripts needed to run `make prepare`.

For compatibility with solutions like DKMS, which do not expect to
run anything but the module build, we add a minimal prepare target
to this path so that the host programs will be automatically rebuilt.
We also make some edits and exclude some files to avoid dependencies
on bison, flex, and OpenSSL.

Signed-off-by: Ben Cressey <bcressey@amazon.com>