packages: express installation dependencies in cargo
Previously we expressed RPM BuildRequires dependencies in Cargo's
dependency graph to ensure the necessary RPM's exist before we build
a package. If we add to this RPM Requires dependencies (dependencies
that are needed when software packages are installed), then we can use
Cargo to select only the packages that are needed for a set of desired
install packages.