os: Add kdump support
Kdump is a Linux feature that allows to boot to a kernel whenever the
system panics. The crash kernel is loaded into a reserved space in
memory determined by the `crashkernel` kernel parameter. In
Bottlerocket, this parameter is set such that no memory will be reserved
if the host has less than 2GB of memory.

For Bottlerocket, the crash kernel is loaded from the current active
boot partition. The `configure-boot-mount.service` systemd unit
determines which is the current active boot partition and mounts it at
`/boot`. This mount is set as read-only and with private propagations,
so new mount namespaces won't have access to it. As part of this change,
SELinux labels are added to the boot partition when it is created by the
`rpm2img` tool.

The `load-crash-kernel.service` systemd unit loads the crash kernel,
only if memory was reserved for it, and the `kexec.kexec_load_disable`
setting is `0`. The unit will exit gracefully if no memory was reserved
for the crash kernel. For the moment only the aws-dev and vmware variants
use that kernel parameter.

The `kexec.kexec_load_disable` setting used to be set in the
`sysctl.conf` configuration file. With this change, the setting is set
using the `disable-kexec-load.service` systemd unit. This unit runs after
`load-crash-kernel.service`, even if the latter wasn't executed or
excited with a non-zero code.

The `capture-kernel-dump.service` systemd unit is set as the target when
the crash kernel is executed. It captures both the dmesg logs and the
kdump-compressed dump excluding:
  * Pages filled with zero
  * Non-private cache pages
  * All cache pages
  * User process data pages
  * Free pages

All the files generated by the `capture-kernel-dump.service` unit are
stored at `/var/log/kdump`, therefore the unit has a strong dependency
on the following services to set up the persistent partition:
  * local-fs.target
  * systemd-sysusers.service
  * systemd-udevd.service
  * systemd-udev-trigger.service
  * systemd-tmpfiles-setup.service
  * systemd-tmpfiles-setup-dev.service

Since `local-fs.target` is a dependency of `capture-kernel-dump.service`,
systemd will attempt to load all the mount units. To prevent this, the
mount units will only be loaded during the execution of the `preconfigured`
target.

No API is provided to enable/disable the dump collection, since the
memory space is reserved and it will be a waste if nothing uses that
space. Dynamically changing the `crashkernel` cmd line parameter isn't
an option since we will provide support for secure boot in the future.