grub: update to 2.06
This moves to Amazon Linux as our upstream, by way of Fedora.

Fedora's patches include an implementation of Secure Boot support, as
well as many other fixes that are not yet upstream.

Amazon Linux's patches include fixes for console handling in EC2, and
on metal instances in particular. They also add an option to disable
module loading, which eliminates a potential vector for altering the
boot process.

Signed-off-by: Ben Cressey <bcressey@amazon.com>