create hybrid images for x86_64
Instead of a single "firmware" partition for either BIOS or EFI, we
now create the BIOS partition for all architectures, and ignore it
except for the x86_64 use case.

Two EFI partitions are allocated out of previously reserved space.
Only the first one is used today; using the second one would require
a scheme to swap the partition types on upgrade, since EFI firmware
will not understand the significance of the GPT priority bits.

The result is a hybrid x86_64 image capable of booting under legacy
BIOS or EFI. With BIOS, the MBR will point to the next GRUB stage in
the BIOS boot partition. With EFI, the firmware will look for the EFI
system partition. In either case, the GRUB image will find the right
boot partition by checking GPT priorities, and load the configuration
file from there.

Signed-off-by: Ben Cressey <bcressey@amazon.com>