pubsys: raise messages to 'warn' if AMI exists or repo doesn't
During development, if you make changes without committing, pubsys will
see that an AMI exists for your (-dirty) commit and not register a new
one.  Similarly, if you have a repo configured in Infra.toml and expect
to be updating it, but it doesn't exist, pubsys will create a new one.
You may not want to continue testing in these cases if you thought a new
AMI was going to be built, or an existing repo updated, with your recent
changes.  Raise these to 'warn' level so they're more obvious.