build: refactor partition layout code
Move all decisions about partition sizes, types, etc into a library
so that it's easier to support layouts for different disk sizes. The
layout is still essentially static, but can be scaled up as needed to
support larger root filesystems.

For the 2 GiB OS image layout, the end result is almost identical,
except for fixing a longstanding bug where the last two partitions on
each disk - PRIVATE and DATA - were (1 MiB + 1 sector) in size, and
not 1 MiB as intended.

Signed-off-by: Ben Cressey <bcressey@amazon.com>