build: add support for building unified images
For some targets such as bare metal systems, the requirement for a
separate block device to hold the data partition is unworkable.

Implement a "unified" image layout, which places the data partition
after the final OS partition, and is suitable for targets which may
only have one disk.

The old "split" layout remains the default.

Signed-off-by: Ben Cressey <bcressey@amazon.com>