packages: add NVIDIA tesla 470 driver
This commit adds the NVIDIA tesla 470 driver. The drivers are
subpackages of `kmod-5.10-nvidia`, since we don't want to have a spec
file per driver type per driver version. The spec file for
`kmod-5.10-nvidia` holds the drivers that are compatible with the 5.10
kernel. New spec files should be added for newer kernel and driver
versions.

The `kmod-5.10-nvidia` package provides a tmpfilesd conf file to create
the lib modules directory where the kernel modules will be created.

Each subpackage installs the libraries and binaries underneath
`%{_cross_bindir}/nvidia/<type>` and `%{_cross_libdir}/nvidia/<type>`
respectively to prevent collisions while building the subpackages.
Kernel module objects are installed in
`%{_cross_datadir}/nvidia/<type>/modules`, so that `driverdog` can
compile them at runtime.

Each subpackage provides a drop-in configuration file for containerd,
that sets the `NVIDIA_PATH` environment variable. This environment
variable must be set to the directory that contains the NVIDIA userland
tools, which will be mounted on the containers by `libnvidia-container`.
The environment variable is set for containerd, since
`libnvidia-container` is called by the runtime to set up the containers.

Signed-off-by: Arnaldo Garcia Rincon <agarrcia@amazon.com>