kernel 5.10: Add config to support additional hardware
This commit adds new config to kernel 5.10 for variants meant to run on bare
metal.  It adds support for SATA drives, Intel networking hardware (1G/10G), as
well as additional suport for UEFI video.  This config will continue to grow
and morph over time as we support additional hardware, etc.