grub: Set `private` variable via search
We currently carry a few patches to enable `search` by partition label
and uuid, but don't include the `search` module.  This commit adds the
module to enable this functionality in our grub config.

This commit also adds an additional command to our grub configs.  The
command uses the `search` module to set the `private` variable to the
location of the `BOTTLEROCKET-PRIVATE` partition.  The `private`
variable will be used to populate the grub menuentry that points to the
bootconfig file.