build: ignore zero blocks on dm-verity root
We omit 512 KiB blocks of all zeroes when registering EBS snapshots.
If an encrypted EBS volume is subsequently created from the snapshot,
reads from these uninitialized regions will return the decryption of
zero with the volume key. This is essentially random garbage and the
hash will not match what dm-verity expects.

These large zero blocks are mostly found in the space between the end
of the filesystem and the partition boundary, but there are some at
the start of the filesystem as well. The regions are not accessed
under normal circumstances. However, tools such as `blkid` will scan
them to fingerprint the device, at which point dm-verity detects the
"corruption" and reboots the instance.

To avoid this, pass the dm-verity option to ignore zero blocks so
that they are never actually read from the backing storage.

Signed-off-by: Ben Cressey <bcressey@amazon.com>