Allow cluster-dns-ip to be a string or list
While clusterDNS is modeled as a list in the kubelet config, we allow
the settings to be given as either a string or list in the API for
backwards compatibility. The API will store and return the data
as-provided, and the model provides a unified iterator view over both
models.