kernel: disable unused filesystems
It's not possible to rule out the existence of workloads using these
filesystems, but it is possible to make a series of educated guesses.

For Kubernetes variants, a CSI driver that supports the filesystem
would be required to use it for container storage. This is especially
true for network-based fileystems, because Bottlerocket does not ship
any of the userspace tools required.

Disabled network filesystems:
* afs - network-based, no CSI driver
* gfs2 - network-based, no CSI driver
* nfs v2 - obsoleted by later versions of NFS

Another use case would be containers that run with CAP_SYS_ADMIN and
mount full disk or filesystem images. Disabling these filesystems is
more of a judgment call, and comes down to whether the format is
obsolete, whether it's in common use, whether it's useful on current
platforms, and if it's consistently enabled across architectures.

Obsolete local filesystems:
* cramfs - read-only format, obsoleted by squashfs
* ecryptfs - obsoleted by native filesystem encryption
* ext2 - obsolete, handled by the ext4 driver
* ext3 - obsolete, handled by the ext4 driver
* romfs - obsoleted by initramfs

Uncommon local filesystems:
* hfs, hfsplus - not enabled on aarch64
* jfs - not enabled on aarch64
* jffs2 - not supported by current platforms
* nilfs2 - not enabled on aarch64
* ntfs - not enabled on 5.10 kernels
* ufs - not enabled on aarch64
* zonefs - not supported by current platforms

Note that a potential use case for hfsplus could be to generate DMG
files for OS X software installs. However, the more common approach
appears to be using `genisoimage` on Linux.

Signed-off-by: Ben Cressey <bcressey@amazon.com>